HNG-STAGE1-TASK/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==1.0.0
Django==4.0.6
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.7.1
gunicorn==20.1.0
idna==3.3
install==1.3.5
oauthlib==3.2.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.11
virtualenv==20.15.1
whitenoise==6.2.0
